requirements.txt
aiohttp==3.8.4
click==8.0.4
gql==3.4.0
httpx==0.23.3
jupyter-book==0.13.1
matplotlib==3.6.2
openpyxl==3.0.10
pandas==1.5.2
python-dotenv==0.21.0
python-dateutil==2.8.2
tenacity==8.2.2
eth-utils==2.1.0
